+++ sw.js
@@ -19,7 +19,7 @@
  */
 'use strict';
 
-var VERSION = 'v114';
+var VERSION = 'v115';
 var RUNTIME_CACHE_VERSION = 'v2';
 var SHELL_CACHE = 'stormscope-shell-' + VERSION;
 var TILE_CACHE = 'stormscope-tiles-' + RUNTIME_CACHE_VERSION;
@@ -69,6 +69,7 @@ var SHELL_ASSETS = [
   './js/surface-observations.js',
   './js/tropical-cyclones.js',
   './js/flood-outlooks.js',
+  './js/winter-outlooks.js',
   './js/river-gauges.js',
   './js/convective-outlooks.js',
   './js/fire-weather.js',